ssl: Certificates and PEM-cache cleaning fixed to avoid memory leak
Certificate db cleaning messages where sent to the wrong process after
restructuring to avoid bottlenecks.

It is possible that the ssl manager process gets two cleaning messages
for the same entry. E.i. first cleaning message is sent and before it
is processed a new reference is allocated  and again released for the
entry, generating a second cleaning message.

Also in ssl_manger:handle_info/2 it is possible that there exists a
new reference to an "old" file name with a potential new content.